Removed Dependency on Seeding Pt.1
