--- .nyc_output/processinfo/a9f2ded0-fbcd-4598-b371-f3db4b358abd.json
@@ -1,1 +1,0 @@
-{"parent":null,"pid":12519,"argv":["/usr/local/bin/node","/mnt/c/Users/Owner/Documents/GitHub/ushift/Product-Information-Service/node_modules/.bin/mocha","_test_/server.spec/server.spec.js","--exit"],"execArgv":[],"cwd":"/mnt/c/Users/Owner/Documents/GitHub/ushift/Product-Information-Service","time":1603248007366,"ppid":12508,"coverageFilename":"/mnt/c/Users/Owner/Documents/GitHub/ushift/Product-Information-Service/.nyc_output/a9f2ded0-fbcd-4598-b371-f3db4b358abd.json","externalId":"","uuid":"a9f2ded0-fbcd-4598-b371-f3db4b358abd","files":["/mnt/c/Users/Owner/Documents/GitHub/ushift/Product-Information-Service/_test_/server.spec/server.spec.js","/mnt/c/Users/Owner/Documents/GitHub/ushift/Product-Information-Service/server/index.js","/mnt/c/Users/Owner/Documents/GitHub/ushift/Product-Information-Service/server/routes/index.js","/mnt/c/Users/Owner/Documents/GitHub/ushift/Product-Information-Service/server/controllers/index.js","/mnt/c/Users/Owner/Documents/GitHub/ushift/Product-Information-Service/database/models/product.js","/mnt/c/Users/Owner/Documents/GitHub/ushift/Product-Information-Service/public/helper-functions.js"]}
